backend/django_test_app/requirements.txt
celery==5.3.1
Django==4.2.4
django-rest-framework==0.1.0
flower==2.0.1
requests==2.31.0
psycopg2-binary==2.9.6
django-cors-headers==3.14.0
